requirements.txt
discord.py
python-dotenv
requests
yfinance
matplotlib
pillow
newspaper3k
feedparser